Change interface of ws_errno_tostr() to return memory.
This change is actually a fix for the function which creates a
ws_value_string object containing a ws_string with the error string, as
whenever this object gets cleaned up, it calls free() on the string,
which should not happen.

It is also a fix for future stuff to be able to construct more advanced
error strings.